+++ doc/ChangeLog.md
@@ -5,6 +5,7 @@
 - Admin.cgi tells jobengine when to reload ftpuploader or force completion of a day.
 - Admin.cgi refuses to enable a ftpuploader rule if directory does not exist
 - Improved signal handling in ftpuploader
+- Make it possible to override default global contants (BaseConfig.pm) using /usr/local/etc/gorm.conf
 
 ## 2019-11-30	sjm	Added partial port of admin.cgi
 - admin.cgi handles create/edit sites, destinations, localdirs. Also forget DOY and finish imcompletes.